iOS-#96 [타이머설정 ViewController] 기능별로 Swift 파일 분리 - layout, Notification, NumericKeyboard, TimerState
